1.32.37
1.32.37:
- Fixed a bug where the event to restore the Byzantine Empire would be
firing repeatedly when it was definitely not appropriate for it to be
firing at all.
- Fixed a bug where a small Byzantium or Latin Empire could be destroyed
even if they controlled Constantinople.

1.32.36:
- Reduced prestige effects of marriage until the 1.09 patch since the AI
does not play well with it presently.
- Tweaked tax income buildings to make money less abundant in the late
game.
- Added a line of cheap tech buildings for castles.

1.32.35:
- Improved AI logic for distributing duchies so that they will hold onto
duchies they have demesne provinces in as long as they are not over the
duchy title limit.
- The Latin Empire and Byzantium will now always be destroyed if
Constantinople is lost to them, regardless of realm size.
- Added a 'Byzantine Restoration' CB against the holder of
Constantinople for independent Orthodox Byzantine culture group rulers
if Byzantium does not exist. A successful war will result in taking the
Duchy of Thrace and recreating the Byzantine Empire.
- Increased prestige impact of marriages to planned 1.09 levels.
- AI now less inclined to wage excommunication wars unless they are
Zealous, have high piety or really dislike the target.
- Issuing a repentance of sins is no longer guaranteed to lift your
excommunication, the Pope may say no depending on opinion and traits. If
the Pope says no you must wait five years to try again.
- An AI ruler whose Empire is about to disband due to lack of holdings
will now attempt to create a Kingdom they are eligible to create in
order to preserve their Duke-level vassals.
- The Golden Horde and Il-Khanate can now end up disbanding their
Empires if they fall below 100 holdings after converting to another
faith.
- The Golden Horde and Il-Khanate will no longer convert to another
faith until they reach at least 100 holdings in size.
- Just about everything that gives you cause to revoke someone's title
now also gives you cause to execute them - but don't expect their
immediate family to be happy about it.
- Faction mood changes due to events and decisions should now be applied
within a few days of the event, instead of at the next maintenance
pulse.
- Grand Tournament now requires you to have reigned for 5 years, to make
it less of a faction pacifier on succession.
- Shattered World easter egg dynasties will no longer show up in the
regular scenario.
- Succession laws of secondary titles will now always be set to the same
as the primary title, even for Kingdoms and above.
- The AI will now change gender succession laws when appropriate.
- Vassal opinion from raised levies will now drop faster, to correspond
with the faster pace of wars in CK2+.
- Reworked the way succession law changes work to prevent any
unintentional loss of prestige or opinion impacts when creating titles.
This is purely a backend change and players should not notice anything
being different.
- Updated to the latest version of Culturally Different Cities.
- Vassals who choose to keep the lands gained in a Crusade will now
relinquish all their old titles to their liege.
- Choosing to keep the lands gained in a Crusade for yourself will now
give you the relevant titles and holy truce modifier, just as if you
gave it to a relative.
- All feudal lords with some participation score in a successful Crusade
now have the option to request a county in the captured Crusade lands
for a landless relative who is not their primary heir. Any characters
granted counties this way will get a large opinion boost with their new
liege.
- Crusader states will now start with High CA, both to make them
stronger and to prevent their lands from being inherited back.
- Fixed various minor bugs and typos.