2.04.0:schwarherz:4 religions abound!
fixed the centering of the religious icons. changed the localization on
reformed norse paganism. Made the "no christians" requirement on the
German Empire simpler (reasoning: if you can choose a highly restricted
HRE or a restrictionless German Empire, why would you pick the HRE?
Also, non-christians needed a creatable empire in that area.)