2.04.0:schwarherz:5 Temporary Removal Alt Scens
Took out the Alt Scenarios in order to combine them into one mod.
Removed the flag=shattered_world lines (or blocks if it was only for
that) from events etc. New Shattered World will be simpler and use the
same modifiers as the rest of the mod for simplicity in updating.