2.04.0:minrog:10 -ancrel namespace, event bug fix
I decided to do the pull request.  I found the event bug in line 531 of
the druidic_holy_order_event.txt file.  There is an OR = { statement
inside of a create_character statement, which caused the next event in
the file to have extra brackets or something.  For now I commented out
the OR = {  and the second option (celtic_pagan) as I wasn't sure what
it was supposed to be.
I also got rid of the Ancient Religions Namespaces.