2.04.0:minrog:18 fix for #119
Al Layun and Abu Khatar had mismatched localisations which was causing
two names to display on top of each other when in county view.
c_el_aaiun had a different main localisation in Kingdoms.txt than it did
in the 0_provinces.txt localisation file.  I changed the localisation in
0_provinces.txt to reflect the name that was in the Kingdoms.txt file.
Also, moved capital of c_idjil in positions.txt so it was inside of its
own county.