2.04.0:schwarherz:34 Karling Nerf + Historical Fix
Removed the k_bavaria from landed titles, gave its duchies to Germany
(it's historical, even in 867). Made the Karlings use Lotharangia over
burgundy by making burgundy have no holder in 867. Made anything that
had k_bavaria as a liege have d_bavaria instead in 867 (increased
instability, also historically accurate). Made the HRE not have de jure
borders in 867 (and require the person hold one of a certain set of
titles in order to form it so we don't get an HRE in the wrong location)