2.05.0:minrog:3 Hashashin bug
The founding event for the Hashashin order was spamming the players at
the beginning of 1089 with the creation message.  The event was changed
from an Immediate + Option to a double-option, but that meant the
set_global_flag wasn't being run if the second option was used.
Probably needs to be changed back to immediate or all pushed into one
option, but for now I added the global flag into the second option
component.