2.05.1:schwarherz:1 religions etc
split the religions into separate files for easier future editing. Gave
the Kemetic names to the head of kemetic religion for reginal names. The
etc is that I added a decision to convert to icelandic culture from
norse. (the AI weight to do it might need tweaked a bit. As it is they
have very little incentive to stay norse unless they're under a norse
emperor)