2.05.3:nivve:9 Alt bookmarks update
Bookmarks created:
- Shatt World 867/1066
- Reign of the Ancients 867 and 867 Shatt world version
- Reign of Princes 867/1066

It's probably still riddled with inconsistencies, but I haven't had time
to debug. Wanted to get it out there for feedback, but if you guys can
find the cause of the 3 digit duplicate ids int he error log, that would
be great (couldn't find those id's in titles and character histories).