2.05.4:minrog:11 Fix for Lake Onega
Made Lake Onega not a sea zone so the Republics wouldn't build trade
posts there.  Lakes must be flagged as "major rivers" in the default.map
file for them to not have ports.  Also changed it so the name would be
"Lake Onega" to differentiate it from the County of Onega in the game
files.