2.06.3:Battle bunny:1 Proselytization fixes
Modified job_lord_spiritual.txt in two ways:
- liege opinion modifiers moved to the end of the MTTH block. I read
somewhere on the forum that the game runs faster if events are
structured like that.

- Fixed the modifiers for MA: previously, the MTTH increased by a factor
of 1.2 for 60+ MA, 1.1 for 70+, and only decreased later on, so 60% MA
is worse than 70%, etc. This has been changed so that the MTTH is
increased and decreased logically.