2.06.3:minrog:17 tech fix, minor stuff
- Fixed Technology at later starting dates.  Due to the end date of
100,000 (which is more likely 32,767), the end_ tech levels need to be
appropriately boosted to make the ratio realistic for our starting
dates.  Probably needs a bit of tweaking on the numbers for each
culture; I tried to roughly match the values that Vanilla was using for
each category in the different dates.

- Changed Robert of Apulia 1066 to have less Piety, but hopefully enough
to fight for d_sicily before he dies

- Fixed some things with c_aprutium

- Enabled Medium Crown Authority for Merchant Republics.  First stage of
Republic makeover