2.07.0:nivve:11 Altbook - disabled AND
Another New Dawn bookmark has issues. Sinc eit is also rather
ocmplicated to maintain as in integral part, I decided it is best to
just leave AND for the standalone ck2+ compatible mod and remove it from
the altbook module. I will remove all the AND files and update all the
other files in due time. I had to go by all files for validator reasons
anyway, so this is a nice excuse to start with that.