2.07.1:nivve:2 female flavour event localisation
Not sure whether the descriptions are accurate, but in the end I didn't
really know what the events were supposed to mimick, so I just made an
educated guess and made some loc for the Martial lady (tengri) and
shield maiden trait events.
If the loc is not correct, feel free to alter it. I don't even know what
the differences between the three traits should be (aside from the
martial point shield maiden gives extra).