2.07.2:Rylock:5
all of the de jure changes are complete, save for east Africa
(Abyssinia, Nubia, Ifar). In order to do those, a complete re-work of
the map is necessary, as the province layout is just too different.

Did a bit of re-working to eastern Russia, but it seems the situation
there is much the same as in east Africa--the vanilla map layout is just
too different compared to the CK2+ map for anything to hold up.

As you can see, the situation is the same in central Asia, the area
between Persia and India. Some re-working of the de jure areas is going
to be necessary for any of it to make sense.