2.07.2:Rylock:65 gone_pagan fix
This fix probably didn't need to be complex as it was--I broke the
decision & event into separate ones for each pagan religion, just to see
if that was the issue, and then discovered the issue was Catholic MA
dropping below 10 and causing every AI character in the game to fire the
decision. Hence overload and CTD.

So probably all that was needed was the AI event which allows them to
use the decision. Likely a better version of this should be made later
which uses something like the SoA system of heretic conversion.

For now, however, this solves the issue of the gone_pagan crash (if not
the massive MA loss, and thus the reason behind it).